This are my week 3 assessments
